chore: move components tp container from pages directory
